Update autoload.php
Remove hard coded -7 from substr.  This is called from Module.php and Manager.php.  Module has only 6 letters which results in an extra \ on the end of the module name for Module.php, and the correct name for Manager.php.  In certain installations this results in include_once trying to include the same file multiple times